Refactor Action and handle GHE Url
When we instantiate the GH Client we now need to take the GHE Url into
consideration, so we moved that out to a builder.

While playing with the new settings we notices that the plugin did not
reload after receiving new settings and some of the behaviour when
moving around was awkward and breaking. So we moved some methods around
and made sure to re-trigger client setup and re-start polling when new
settings are received. This also allowed moving the communication of
status to a more relevant place.